requirements-databricks.txt
# Databricks-compatible requirements for ENTSOE ETL pipeline
# These versions are tested with Databricks Runtime 13.3 LTS

# Core dependencies
requests>=2.28.0,<3.0.0
xmltodict>=0.13.0,<1.0.0
pandas>=1.5.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pytz>=2022.0,<2024.0
pydantic>=1.10.0,<3.0.0
tenacity>=8.0.0,<9.0.0

# Optional: SQLAlchemy for advanced database operations
sqlalchemy>=1.4.0,<3.0.0

# Optional: Additional utilities for Databricks
# These are often pre-installed in Databricks but included for completeness
numpy>=1.21.0
pyarrow>=10.0.0 